eyeball: Remove shared::Observable::try_into_unique
It didn't actually work as the documentation said, and it was not
included in any releases. Specifically, it didn't work when there were
any subscribers. It would be sound to allow that, but would require
extra unsafe code that I don't want to add right now, before there's
even a use case for the method.