feat(im-util) Implement `DynamicLimit`.
A `VectorDiff` stream adapter that presents a limited view of the
underlying `ObservableVector`s items.

For example, be `S` a `Stream<Item = VectorDiff>`. The `Vector` represented
by `S` can have any length, but one may want to virtually _limit_ this `Vector`
to a certain size. Then this `DynamicLimit` adapter is well appropriated.
The limit is dynamic, i.e. it changes over time based on values that are polled
from another `Stream` (ref. `Self::limit_stream`).

Because the limit is dynamic, an internal buffered vector is kept, so that
the adapter knows which values can be added when the limit is increased, or
when values are removed and new values must be inserted. This fact is important
if the items of the `Vector` have a non-negligeable size.

It's OK to have a limit larger than the length of the observed `Vector`.

Note that this adapter won't produce anything until a new limit is polled.