feat: Replace `arrayvec` by `smallvec`.
The motivation behind this patch is to replace `arrayvec` by `smallvec`.
Why? Because `arrayvec` cannot grow, it's an array-baked vector. Once
the full capacity is reached, it cannot grow any more. That was fine
up until now because it was only used by the `Limit` stream adapter
which needs only 2 values in a buffer that is typed by an `ArrayVec`.
But for other stream adapters, it's a limitation. What if a stream
adapter needs a buffer of more than 2 values?

Ideally, we want to preserve the stack-allocated property for obvious
performance reasons (cache locality, reducing allocator traffic and so
on). However, we also want to be able to fallback to the heap for larger
allocations.

`smallvec` provides this feature. Hence, it does what `arrayvec` does,
with the addition of heap-allocated fallback when it's needed.