feat(im): Add support for `serde`.
This patch adds the `serde` feature, which —when turned on— implements
`serde::Serialize` for `VectorDiff`.

This is a custom `serde::Serialize` implementation because the `derive`
macro cannot be used here, since it requires `T` in `VectorDiff<T>` to
implement `serde::Serialize` and `Clone`. These constraints must not be
present on the `VectorDiff` enum definition.

This patch also adds a test suite: it tests that all `VectorDiff`'s
variants can be serialized to JSON for example.

Note that `VectorDiff` doesn't implement `serde::Deserialize` because
it's not necessary: `eyeball-im` only _emits_ `VectorDiff`s, it cannot
receive them.