chore: Rename `Limit` to `Head`.
This patch renames the higher-order stream `Limit` to `Head`. It is
a reaction to the upcoming https://github.com/jplatte/eyeball/pull/69
where I've initially proposed the name `RLimit`. After a short thinking,
I believe renaming `Limit` to `Head` and `RLimit` to `Tail` would be
better names.

This patch renames `Limit` to `Head`.