fix(util): Fix when `SortBy<Stream<Item = T>>` handles a `VectorDiff::Set` where `T` is a shallow clone type.
So. When `SortBy` was handling a `VectorDiff::Set`, it was finding
the position of the old value, the position of the new value, and was
trying to figure out where to insert the new value. Removing the old
value isn't done every time, but when it was done, it was done after
the position of the new value was calculated. It creates a fundamental
problem: **When 2 values are shallow clones of the same initial value**.

Imagine the type:

```rust
struct T {
    inner: Arc<Mutex<usize>>,
}

let mut ob = ObservableVector::<T>::new();
let (values, mut sub) = ob.subscribe().sort_by(cmp);

let t1 = T { inner: Arc::new(Mutex::new(42)) };
let t2 = t1.clone();

// Insert `t1` at position 0.
v.push_back(t1);

*t2.inner.lock().unwrap() = 43;

// Update value at position 0.
v.set(0, t2);
```

To sort this, we need a _compare function_ like so:

```rust
fn(left: &T, right: &T) -> Ordering
```

In our case, it looks like:

```
fn cmp(left: &T, right: &T) -> Ordering {
    let left = left.inner.lock().unwrap();
    let right = right.inner.lock().unwrap();

    left.cmp(&*right)
}
```

What happens then?

- **Previously** — The _old_ value (the value being replaced) was not
  removed from the buffered vector before computing the new position.
  So, at some point, our compare function will receive the old value
  in `left` and the new value in `right` (or vice versa, but that's not
  important). It's _impossible_ to sort `left` and `right` here, because
  technically, updating `t2` to contain `43` will also update `t1`: `t1`
  no longer contains `42` but also `43`. Boom 💥. It creates an infinite
  loop.
- **Now** — The position of the _old_ value is computed, then the _old_
  value is _removed_ from the buffered vector. Only now, we compute the
  position of the new value. And now, we can insert the new value at the
  correct position and emit `VectorDiff`s.

Pros of the new strategy:

- our compare function no longer receives the _old_ value in `left`,
- it fixes the bug with shallow clones,
- it semantically is more correct,
- it saves one call to our compare function (because there is one less
  value in the vector).

Con of the new strategy:

- if the new value ends up being inserted at the same position of the
  old value, we call `remove` and `insert` instead of `set`, which can
  be less efficient (that's precisely what the previous code was trying
  to avoid, this optimisation is the cause of the bug).

I know `Vector` is quite optimised, which makes feel like the only con
can be considered negligible compared to the pros (which include the
bug fix!).

This patch does that. In addition, funnily, it's easier to optimise the
`VectorDiff::Insert`: it can be mapped to `VectorDiff::PushFront` or
`VectorDiff::PushBack` when applicable.